README.md:fix spelling, add link
His name is Titus (which is public information), and uses the nick
"titi" in the game. His GitHub username is @titiger (titi + "ger", which
stands for German).

I also took the liberty of adding a link to his GitHub profile.

And thank you for mentioning me. Thank you for your hard work
@mathusummut